services/gateway/pyproject.toml
[project]
name = "gateway"
version = "0.1.0"
description = "Gateway service scaffold"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.125.0",
    "langchain>=1.2.0",
    "langchain-community>=0.3.0",
    "grpcio>=1.67.0",
    "grpcio-tools>=1.67.0",
    "sentence-transformers>=3.0.1",
    "pydantic-settings>=2.7.0",
    "uvicorn>=0.32.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 88
src = ["src"]
fix = true
